requirements.txt
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.0.1
streamlit==1.7.0
streamlit_cropper==0.2.0

